Add Bootloader Assert
Guard against users with 5.1 bootloader to be safe.

Change this for 12.1

Change-Id: I332a5cd81a8ecf4a28fd278d448bb1a9a6c44158

Expand assert

Users who take the 5.1 OTA are otherwise stuck with the 7108
bootloader. Attempting to downgrade forceably will result in a brick
after an unknown number of attempts. Fastboot downgrade is likewise
blocked by a prevalidation check.

Allow nightlies to be used against either sets of firmware.

Change-Id: I88b177a2defd8ce93d4edd11c57255584951586a