Update patch set 2
Patch Set 2:

I think the reason you are getting a black screen is because SELinux is enforcing and init is refusing to start the recovery service because I don't have an selabel set for the recovery service in order to maintain compatibility with older Android trees. I found a fairly good method for setting SELinux permissive in all recovery builds that will still allow Omni to boot in enforcing. I'll create the patches and push them up to Gerrit tomorrow.

Patch-set: 2