Update patch set 2
Patch Set 2:

Not sure if you understood my comment. In twrp.fstab, in the options for /data, it has listed a flag for "forceencrypt=..."  On the regular boot fstab, this causes the OS to forcibly encrypt the /data partition, so I was asking if TWRP will act the same way.  It might be safer to change that flag to "encryptable=..."

Patch-set: 2