Update patch set 2
Patch Set 2:

The support for the forceencrypt flag is more for copy/paste convenience. We do not encrypt the device in recovery for a number of reasons. I pushed these 2 patches up to Gerrit which will allow us to set SELinux into permissive for recovery only:
https://gerrit.omnirom.org/10849 and https://gerrit.omnirom.org/10850

I pushed the changes to init in system/core up to AOSP as well.

Patch-set: 2