Update patch set 1
Patch Set 1:

Yea that's what I thought too, but it does seem that CM was forcing the bootloader (and modem, to my knowledge) in their nightlies (not during compilation, probably something done server-side, because there is no trace of it in their device tree)

We also found out that Google changed the security version, preventing downgrading bootloader altogether.

Ex: you cannot be on a 5.1 bootloader and flash the factory images containing 5.0 bootloader anymore (kind of defeats the purpose of a Nexus device).

This was done (what we believe) by changing the security version.

Patch-set: 1