Shamu: make inheriting vendor makefile optional
For the people who just want to compile TWRP, lunching with just the device tree
and kernel source will result in an error due to missing vendor files. This line
has to be deleted or changed to this in order to proceed. Mako and Hammerhead
have it optional so why not here as well?

Change-Id: I10f73d6b8d6655476a38613f06b5481eb1faecd0